requirements.txt
pytest>=8.2.0
selene>=2.0.0rc4
requests==2.32.3
allure-pytest>=2.13.5
appium-python-client==4.0.0
pydantic>=2.7.4
python-dotenv==1.0.1